+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added support for reading and writing `.wilut` files to to xc3_tex.
+
 ## 0.19.0 - 2025-04-03
 ### Added
 * Added support for wimdo models from Xenoblade X DE.